refactor: update MongoDbExtensions to have a common place for adapter configuration
